BUG: Print-Shop Page Images
Gatsby-node source shopify updated to
(a): add the first product img as the localFile associated
with the product;
(b): add the variant images as the localFile associated with
the product variants
(c): any other product images are not directly associated with
the product, but are linked to the product via a foreign key
named "parentNodeId".

Also, if the first product image IS the first and only variant
image, then don't add it and just rely on the "Default Title"
variant.

This is semi complex and we might need to come back to this.